fixed some values
git-svn-id: https://svn.vis.ethz.ch/svn/kob_gamelab09/trunk@193 fc16ba98-2866-4055-be8f-df951a5456ab

+++ src/ProjectMagma/ProjectMagma/Framework/Properties/PlayerControllerProperty.cs
@@ -120,10 +120,10 @@ namespace ProjectMagma.Framework
                     fuel -= gameTime.ElapsedGameTime.Milliseconds;
                     jetpackVelocity += jetpackAcceleration * dt;
 
-                    if (jetpackVelocity.Length() > constants.GetInt("max_jetpack_speed"))
+                    if (jetpackVelocity.Length() > constants.GetFloat("max_jetpack_speed"))
                     {
                         jetpackVelocity.Normalize();
-                        jetpackVelocity *= constants.GetInt("max_jetpack_speed");
+                        jetpackVelocity *= constants.GetFloat("max_jetpack_speed");
                     }
                 }
             }
